facebook-fetcher/requirements.txt
arrow==0.10.0
certifi==2017.7.27.1
chardet==3.0.4
facebook-sdk==2.0.0
idna==2.6
pika==0.11.0
pymongo==3.5.1
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
urllib3==1.22
